+++ playroom/snippets.tsx
@@ -10,7 +10,7 @@ const menuSnippet = {
     <Menu
       width={400}
       renderTarget={({ ref, onPress, isMenuOpen }) => (
-        <Touchable elementRef={ref} onPress={onPress} style={{ width: 100 }}>
+        <Touchable ref={ref} onPress={onPress} style={{ width: 100 }}>
           <Inline space={16}>
             <IconKebabMenuLight />
             <Text3 regular>{isMenuOpen ? "Close" : "Open"}</Text3>
@@ -628,16 +628,40 @@ const cardSnippets: Array<Snippet> = [
             buttonLink={<ButtonLink onPress={() => {}}>Link</ButtonLink>}
         />`,
     },
+    {
+        group: 'Cards',
+        name: 'SnapCard',
+        code: `
+        <SnapCard
+            title="Title"
+            subtitle="Subtitle"
+            icon={
+              <Circle size={40} backgroundColor={colors.brand}>
+                  <ThemeVariant isInverse>
+                      <IconAcademicRegular />
+                  </ThemeVariant>
+              </Circle>
+            }
+        />`,
+    },
 ];
 
 const titlesSnippets: Array<Snippet> = [
     {
-        name: 'SectionTitle',
-        code: '<SectionTitle>Some title</SectionTitle>',
+        name: 'Title1',
+        code: '<Title1>Some title</Title1>',
     },
     {
-        name: 'SectionTitle (with link)',
-        code: '<SectionTitle right={<TextLink onPress={() => {}}>Link</TextLink>}>Some title</SectionTitle>',
+        name: 'Title1 (with link)',
+        code: '<Title1 right={<TextLink onPress={() => {}}>Link</TextLink>}>Some title</Title1>',
+    },
+    {
+        name: 'Title2',
+        code: '<Title2>Some title</Title2>',
+    },
+    {
+        name: 'Title2 (with link)',
+        code: '<Title2 right={<TextLink onPress={() => {}}>Link</TextLink>}>Some title</Title2>',
     },
 ].map((snippet) => ({...snippet, group: 'Titles'}));
 
@@ -716,7 +740,7 @@ const layoutSnippets: Array<Snippet> = [
                 },
                 ].map((category) => (
                 <Stack key={category.categoryName} space={8}>
-                    <SectionTitle>{category.categoryName}</SectionTitle>
+                    <Title1>{category.categoryName}</Title1>
                     <NegativeBox>
                     <RowList>
                         {category.settings.map((setting) => (
@@ -863,7 +887,7 @@ const exampleScreens: Array<Snippet> = [
                   },
                 ].map((category) => (
                   <Stack key={category.categoryName} space={8}>
-                    <SectionTitle>{category.categoryName}</SectionTitle>
+                    <Title1>{category.categoryName}</Title1>
                     <NegativeBox left right={!isDesktopOrBigger}>
                       <RowList>
                         {category.settings.map((setting) => (
@@ -900,7 +924,7 @@ const exampleScreens: Array<Snippet> = [
                 </Text3>
               </Stack>
               <Stack space={8}>
-                <SectionTitle>Section 1</SectionTitle>
+                <Title1>Section 1</Title1>
                 <NegativeBox>
                   <RowList>
                     <Row title="Title" switch={{ defaultValue: false }} />
@@ -909,7 +933,7 @@ const exampleScreens: Array<Snippet> = [
                 </NegativeBox>
               </Stack>
               <Stack space={8}>
-                <SectionTitle>Section 2</SectionTitle>
+                <Title1>Section 2</Title1>
                 <NegativeBox>
                   <RowList>
                     <Row
@@ -1140,6 +1164,34 @@ const carouselSnippets = [
     },
 ];
 
+const avatarSnippets = [
+    {
+        group: 'Avatar',
+        name: 'Avatar with badge',
+        code: `<Avatar size={64} src="https://i.imgur.com/nRBEMMV.png" badge />`,
+    },
+    {
+        group: 'Avatar',
+        name: 'Avatar with numeric badge',
+        code: `<Avatar size={64} src="https://i.imgur.com/nRBEMMV.png" badge={5} />`,
+    },
+    {
+        group: 'Avatar',
+        name: 'Avatar with initials',
+        code: `<Avatar size={64} initials="PL" />`,
+    },
+    {
+        group: 'Avatar',
+        name: 'Avatar with icon',
+        code: `<Avatar size={64} />`,
+    },
+    {
+        group: 'Avatar',
+        name: 'Avatar with custom icon',
+        code: `<Avatar size={64} Icon={IconFireRegular} />`,
+    },
+];
+
 export default [
     ...buttonSnippets,
     ...formSnippets,
@@ -1266,4 +1318,5 @@ export default [
         code: `<Image src="https://i.imgur.com/aEVtKsE.jpg" aspectRatio="16:9" />`,
     },
     ...carouselSnippets,
+    ...avatarSnippets,
 ].sort((s1, s2) => s1.group.localeCompare(s2.group)) as Array<Snippet>;