+++ src/__stories__/centered-carousel-story.tsx
@@ -12,7 +12,7 @@ import {
     IconInformationRegular,
     Text10,
     ThemeVariant,
-    SectionTitle,
+    Title1,
 } from '..';
 
 export default {
@@ -70,7 +70,7 @@ export const Default: StoryComponent<Args> = ({numItems}) => {
                         ))}
                     />
                     <Stack space={8}>
-                        <SectionTitle as="h2">Current page info</SectionTitle>
+                        <Title1 as="h2">Current page info</Title1>
                         {pageInfo && (
                             <Text2 regular as="pre">
                                 {JSON.stringify(pageInfo, null, 2)}