+++ src/__stories__/field-story.tsx
@@ -20,11 +20,11 @@ import {
     Text1,
     ButtonPrimary,
     Form,
-    SectionTitle,
+    Title1,
     Stack,
 } from '..';
 import {inspect} from 'util';
-import Icon from '../icons/icon-visibility';
+import IconMusicRegular from '../generated/mistica-icons/icon-music-regular';
 import {StorySection, countriesList, phoneNumbersList} from './helpers';
 import {getLocalDateString, getLocalDateTimeString} from '../utils/time';
 
@@ -188,7 +188,12 @@ export const Variants: StoryComponent = () => {
             </StorySection>
 
             <StorySection title="With icon at the end">
-                <TextField name="text" label="Label" defaultValue="Default value" endIcon={<Icon />} />
+                <TextField
+                    name="text"
+                    label="Label"
+                    defaultValue="Default value"
+                    endIcon={<IconMusicRegular />}
+                />
             </StorySection>
 
             <StorySection title="Disabled">
@@ -199,7 +204,7 @@ export const Variants: StoryComponent = () => {
                         disabled
                         label="Disabled"
                         defaultValue="Default value"
-                        endIcon={<Icon />}
+                        endIcon={<IconMusicRegular />}
                     />
                     <SearchField name="search" disabled label="Search" />
                     <PhoneNumberField
@@ -704,7 +709,7 @@ export const DateTimeLimits: StoryComponent = () => {
 
     return (
         <Form onSubmit={() => alert({message: 'Success!'})}>
-            <SectionTitle>Date Field with range</SectionTitle>
+            <Title1>Date Field with range</Title1>
             <Stack space={16}>
                 <DateField
                     name="date"
@@ -717,7 +722,7 @@ export const DateTimeLimits: StoryComponent = () => {
                 <Text1 regular>Selected value: {dateValue}</Text1>
             </Stack>
 
-            <SectionTitle>DateTime Field with range</SectionTitle>
+            <Title1>DateTime Field with range</Title1>
             <Stack space={16}>
                 <DateTimeField
                     name="datetime"